refactor(linter/plugins): split TSDown plugins into separate files (#20304)
Pure refactor. TSDown config had got very large with the custom plugins to replace `debugAssert` calls, and hoist globals vars to top level. A PR to follow will make the situation worse, as it adds another plugin.

Move the plugins into separate files, to bring the TSDown config back to a manageable size.

This PR does not alter the plugin code, just moves it.
