+++ apps/oxlint/package.json
@@ -39,6 +39,7 @@
     "@arethetypeswrong/core": "catalog:",
     "@babel/code-frame": "^7.28.6",
     "@napi-rs/cli": "catalog:",
+    "@oxc-project/types": "^0.118.0",
     "@type-challenges/utils": "^0.1.1",
     "@types/babel__code-frame": "^7.27.0",
     "@types/esquery": "^1.5.4",
